requirements.txt
aiohttp==3.12.13
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
faiss-cpu==1.11.0
fastapi==0.115.13
ffmpeg-python==0.2.0
future==1.0.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-python-client==2.173.0
google-auth==2.40.3
google-generativeai==0.8.5
gotrue==2.12.0
grpcio==1.73.0
httpx==0.28.1
Jinja2==3.1.6
jiter==0.10.0
limits==5.4.0
multidict==6.5.0
numpy==2.3.1
openai==1.90.0
packaging==25.0
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.2
PyMuPDF==1.26.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
realtime==2.4.3
requests==2.32.4
rsa==4.9.1
scipy==1.16.0
slowapi==0.1.9
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.3
supafunc==0.9.4
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
websockets==14.2
wrapt==1.17.2
yarl==1.20.1

# Additional dependencies for enhanced functionality
aiofiles==24.1.0
Pillow==11.1.0
python-magic==0.4.27